requirements.txt
asttokens==3.0.1
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.3
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-translate==3.24.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
ipython==8.38.0
jedi==0.19.2
matplotlib-inline==0.2.1
metadata_parser==1.0.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
Pygments==2.19.2
PySocks==1.7.1
python-telegram==0.19.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
stack-data==0.6.3
telegram_text==0.2.0
tldextract==5.3.1
traitlets==5.14.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.3.5
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
